+++ examplepatches/LaKakaReal/LaKakaShitPatches/patches/src/main/kotlin/app/template/patches/instapaper/Fingerprints.kt
@@ -1,0 +1,9 @@
+package app.template.patches.instapaper
+
+import app.morphe.patcher.Fingerprint
+
+object PremiumCheckFingerprint : Fingerprint(
+    strings = listOf("subscription_is_active", "0"),
+    returnType = "Ljava/lang/String;",
+    parameters = listOf()
+)