+++ examplepatches/Xhehab/Xhehab-Patches/patches/src/main/kotlin/app/template/patches/shared/Constants.kt
@@ -20,7 +20,7 @@ object Constants {
         packageName = "uni.cbum",
         apkFileType = ApkFileType.XAPK,
         appIconColor = 0x111111,
-        targets = listOf(AppTarget(version = "4.1.43"))
+        targets = listOf(AppTarget(version = "4.1.44"))
     )
 
     // FST-7
@@ -30,5 +30,14 @@ object Constants {
         apkFileType = ApkFileType.APKS,
         appIconColor = 0xD7262E,
         targets = listOf(AppTarget(version = "4.0.1"))
+    )
+
+    // Video Player All Format
+    val VIDEO_PLAYER_COMPATIBILITY = Compatibility(
+        name = "Video Player All Format",
+        packageName = "video.player.videoplayer",
+        apkFileType = ApkFileType.APKM,
+        appIconColor = 0x18DC14,
+        targets = listOf(AppTarget(version = "2.6.2"))
     )
 }