+++ examplepatches/LaKakaReal/LaKakaShitPatches/CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.3.0](https://github.com/LaKakaReal/LaKakaShitPatches/compare/v1.2.0...v1.3.0) (2026-07-08)
+
+### ✨ New Features
+
+* **Instapaper:** added instapaper ([4050e7a](https://github.com/LaKakaReal/LaKakaShitPatches/commit/4050e7ae6e494c067471692c9ecd12f8ae786e8f))
+* **Soundbrenner:** added soundbrenner ([b98bf61](https://github.com/LaKakaReal/LaKakaShitPatches/commit/b98bf61ffef0e435bd91a74bbabcc72472c96dc3))
+
+## [1.3.0-dev.1](https://github.com/LaKakaReal/LaKakaShitPatches/compare/v1.2.0...v1.3.0-dev.1) (2026-07-08)
+
+### ✨ New Features
+
+* **Instapaper:** added instapaper ([4050e7a](https://github.com/LaKakaReal/LaKakaShitPatches/commit/4050e7ae6e494c067471692c9ecd12f8ae786e8f))
+* **Soundbrenner:** added soundbrenner ([b98bf61](https://github.com/LaKakaReal/LaKakaShitPatches/commit/b98bf61ffef0e435bd91a74bbabcc72472c96dc3))
+
 ## [1.2.0](https://github.com/LaKakaReal/LaKakaShitPatches/compare/v1.1.1...v1.2.0) (2026-06-22)
 
 ### ✨ New Features